Novo template com app.ts
