+++ modulo5/projeto-labook/package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "scripts": {
     "migrations": "tsc && node ./build/migrations.js",
-    "start": "npm run build && node --inspect ./build/src/index.js",
-    "dev": "clear && ts-node-dev ./src/index.ts",
+    "start": "npm run build && node --inspect ./build/src//controller/app.js",
+    "dev": "clear && ts-node-dev ./src/controller/app.ts",
     "build": "clear && echo \"Transpiling files...\" && tsc && echo \"Done!\" ",
     "test": "jest"
   },
