+++ modulo5/projeto-labook/src/controller/UserController.ts
@@ -1,0 +1,7 @@
+/* import { Request, Response } from "express";
+
+export default class userController{
+    signup = async(req: Request, res: Response) => {
+
+    }
+} */
