+++ modulo5/projeto-labook/src/index.ts
@@ -1,19 +1,3 @@
-import dotenv from "dotenv";
-import express, { Express } from "express";
-import cors from "cors";
-import {AddressInfo} from "net";
+import { app } from "./controller/app";
 
-dotenv.config();
-
-const app: Express = express();
-app.use(express.json());
-app.use(cors());
-
-const server = app.listen(3003, () => {
-  if (server) {
-    const adress = server.address() as AddressInfo;
-    console.log("Servidor rodando na porta 3003");
-  } else {
-    console.log("Server is not available");
-  }
-});
+/* app.post("/user/signup", userController.signup); */
