+++ modulo5/projeto-labook/src/types/authData.ts
@@ -1,3 +1,3 @@
-export type AuthenticationData = { 
+export type authenticationData = { 
     id: string
 }
